pydantic gen working but no validators
