+++ examples/string/metadata.yaml
@@ -8,3 +8,5 @@ dataset_description:
 
 source_files:
   - 'protein-links-detailed.yaml'
+  - 'protein-links-detailed-2.yaml'
+
