+++ examples/string/protein-links-detailed.yaml
@@ -4,6 +4,7 @@ delimiter: ' '
 
 files:
   - './examples/data/string.tsv'
+  - './examples/data/string.tsv'
 
 columns:
   - 'protein1'
