+++ koza/model/biolink/__init__.py
@@ -1,4 +1,1 @@
-from .association import *
-from .entity import *
-from .named_thing import *
-from .predicate import predicate
+
